requirements.txt
types-requests
supabase
aiosqlite
python-dotenv
quart
aiohttp